CRM Simulator for banking
For sandbox one follow these steps :
Click on "Click here to start" to initiate the sandbox environment.
In the top-left corner, click on the 4th icon (a profile icon) to access the WhatsApp chat dashboard.
Click on the first chat, and it will open in the center of the next window. There, you can scroll through the chat.
Alternatively, click on the thumbs-up icon to reveal a list of tabs. You can check the "Accounts" and "Transactions" tabs as well.
To return to the chat screen, click on "Chat History."
In the chat window, open the file sent by the customer and note the reference number as "TS123456789."
To check the transaction, click on "Payhub Pro" in the top-right corner.
Click on "Login," and you will be directed to the dashboard page.
From the top-middle section, choose "Reports," then select "Inward Transfer Detail Report."
Choose the message category as "Financial" or any other from the dropdown menu.
Enter the reference number under "Sender Ref No" as "TS123456789."
Click the search button, and you will be able to view the transaction.
At the bottom-right, click on the "Thinkowl" button to return to Thinkowl. Scroll down and review the chat's conclusion.
As a resolution has been provided to the customer, click on "Open" at the top-left corner, and select "Resolved" to close the case.
Once done, you can either restart or close the course, as the simulation of "Inward International Payment" is completed

Important keywords :
For responses : greet, auth, inward( select inward payment not received ), closing
For classification on right of screen search “inward” or use drop down , for left top search for “payment” in place of inbox and select “international payments” then click on unassigned and select name of advisor as “Harshu Dhingra”